Moved transform file from get-weather-data funcion to separate package to reuse in oder functions.
